% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{append_format}
\alias{append_format}
\title{Format string for filename purposes}
\usage{
append_format(s, q, n)
}
\arguments{
\item{s}{String, name of quantity}

\item{q}{Numeric, actual value of quantity}

\item{n}{Positive integer, number of decimal places desired}
}
\description{
Format string for filename purposes
}
