% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{filename_body}
\alias{filename_body}
\title{Construct middle part of filename of simulation run results}
\usage{
filename_body(alpha, theta, n, data.seed, p)
}
\arguments{
\item{alpha}{Real value less than 1, concentration parameter of the Pitman-Yor process}

\item{theta}{Real value, >= -alpha if alpha is between 0 (inclusive) and 1 (exclusive), or a positive integer multiple of |alpha| if alpha < 0}

\item{n}{Positive integer, number of nodes of the network i.e. order of corresponding graph}

\item{data.seed}{Positive integer, the seed for set.seed()}

\item{p}{Real value between 0 and 1 inclusive, prior probability of finite regime}
}
\description{
Construct middle part of filename of simulation run results
}
