% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gvs}
\alias{gvs}
\title{MCMC sampler of DAG-SBM, with possible model selection between finite and infinite regime}
\usage{
gvs(Y, sigma, scalars)
}
\arguments{
\item{Y}{A sparse square matrix}

\item{sigma}{An integer vector that is a permutation of 0, 1, ..., ncol(Y)-1}

\item{scalars}{Data frame of 1 row with the following columns: K, seed, iter, thin, burn, freq, node, scan, L, pg, p_fin, mean_k, a_gamma, b_gamma, a_theta, b_theta, a_alpha, b_alpha, a1, b1, aa0, ba0, ab0, bb0, axi, bxi, ak, bk, dag}
}
\description{
MCMC sampler of DAG-SBM, with possible model selection between finite and infinite regime
}
