% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{obtain_point_est}
\alias{obtain_point_est}
\title{Obtain point estimate of the allocation vector}
\usage{
obtain_point_est(Z.mat, g, ...)
}
\arguments{
\item{Z.mat}{Data frame of MCMC samples of the node memberships}

\item{g}{igraph object of the underlying graph}

\item{...}{other arguments passed to salso::salso()}
}
\description{
Obtain point estimate of the allocation vector
}
