% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{obtain_spi}
\alias{obtain_spi}
\title{Obtain seconds per iteration of MCMC runs}
\usage{
obtain_spi(alpha, theta, n, p, data.seed, dir)
}
\arguments{
\item{alpha}{Real value less than 1, concentration parameter of the Pitman-Yor process}

\item{theta}{Real value, >= -alpha if alpha is between 0 (inclusive) and 1 (exclusive), or a positive integer multiple of |alpha| if alpha < 0}

\item{n}{Positive integer, number of nodes of the network i.e. order of corresponding graph}

\item{p}{Real value between 0 and 1 inclusive, prior probability of finite regime}

\item{data.seed}{Positive integer, the seed for set.seed()}

\item{dir}{Boolean, is the SBM for directed graphs (TRUE) or directed acyclic graphs (FALSE)?}
}
\description{
Obtain seconds per iteration of MCMC runs
}
\seealso{
\code{\link{compute_bf}}, \code{\link{read_results}} and \code{\link{plot_topo_wrapper}} with the same set of arguments
}
