% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{pivot_pars}
\alias{pivot_pars}
\title{Restructure data frame of MCMC samples for pivot_topo() and subsequent plotting}
\usage{
pivot_pars(df, var_names)
}
\arguments{
\item{df}{Data frame of MCMC samples, with at least the following column names: K, k, gamma, theta, alpha, a0, b0, finite}

\item{var_names}{Character vector with one or more of the following values (in quotes): Kn, k, gamma, theta, alpha, log(gamma), a, b}
}
\description{
Restructure data frame of MCMC samples for pivot_topo() and subsequent plotting
}
