% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{pivot_topo}
\alias{pivot_topo}
\title{Restructure data frame for topological ordering plotting}
\usage{
pivot_topo(df, v = NULL)
}
\arguments{
\item{df}{Data frame, normally output of pivot_pars}

\item{v}{Integer vector, a permutation of 1, 2, ..., length(v)}
}
\description{
Restructure data frame for topological ordering plotting
}
