% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_adj_mat}
\alias{plot_adj_mat}
\title{Plot adjacency matrix}
\usage{
plot_adj_mat(df.edgelist, df.topo, n, weighted = TRUE)
}
\arguments{
\item{df.edgelist}{asdf}

\item{df.topo}{asdf}

\item{n}{asdf}

\item{weighted}{asdf}
}
\description{
Plot adjacency matrix
}
