% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_adj_psm}
\alias{plot_adj_psm}
\title{Plot adjacency matrix overlaid by similarity (co-clustering) matrix}
\usage{
plot_adj_psm(l, width)
}
\arguments{
\item{l}{List, usually returned by \code{\link{obtain_point_est}}}

\item{width}{Numeric, line width of the vertical & horizontal lines separating the blocks}
}
\description{
Plot adjacency matrix overlaid by similarity (co-clustering) matrix
}
