% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_adj_wrapper}
\alias{plot_adj_wrapper}
\title{Plot adjacency matrix with the rows and columns in topological order}
\usage{
plot_adj_wrapper(alpha, theta, n, p, data.seed, dir, A_or_Y)
}
\arguments{
\item{alpha}{Real value less than 1, concentration parameter of the Pitman-Yor process}

\item{theta}{Real value, >= -alpha if alpha is between 0 (inclusive) and 1 (exclusive), or a positive integer multiple of |alpha| if alpha < 0}

\item{n}{Positive integer, number of nodes of the network i.e. order of corresponding graph}

\item{p}{Real value between 0 and 1 inclusive, prior probability of finite regime}

\item{data.seed}{Positive integer, the seed for set.seed()}

\item{dir}{Boolean, is the SBM for directed graphs (TRUE) or directed acyclic graphs (FALSE)?}

\item{A_or_Y}{Boolean, is it the average (TRUE) or actual (FALSE) adjacency matrix that is required from the object of \code{\link{read_results}}?}
}
\description{
Plot adjacency matrix with the rows and columns in topological order
}
