% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap}
\usage{
plot_heatmap(alpha, theta, n, p, data.seed, A_or_Y)
}
\arguments{
\item{alpha}{asdf}

\item{theta}{asdf}

\item{n}{asdf}

\item{p}{asdf}

\item{data.seed}{asdf}

\item{A_or_Y}{asdf}
}
\description{
Plot heatmap
}
