% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_subset}
\alias{plot_subset}
\title{Plot subset}
\usage{
plot_subset(
  df,
  alpha,
  theta,
  data.seeds,
  ns = c(250, 500, 1000),
  scales = "free"
)
}
\arguments{
\item{df}{asdf}

\item{alpha}{asdf}

\item{theta}{asdf}

\item{data.seeds}{asdf}

\item{ns}{asdf}

\item{scales}{asdf}
}
\description{
Plot subset
}
