% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_topo}
\alias{plot_topo}
\title{Plot topological ordering}
\usage{
plot_topo(
  df,
  colours = TRUE,
  mean.pos = FALSE,
  low = "yellow",
  true.pos = NULL
)
}
\arguments{
\item{df}{asdf}

\item{colours}{asdf}

\item{mean.pos}{asdf}

\item{low}{asdf}

\item{true.pos}{asdf}
}
\description{
Plot topological ordering
}
