% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_topo_wrapper}
\alias{plot_topo_wrapper}
\title{Wrapper of \code{\link{plot_topo}}}
\usage{
plot_topo_wrapper(alpha, theta, n, p, data.seed, dir = FALSE)
}
\arguments{
\item{alpha}{Real value less than 1, concentration parameter of the Pitman-Yor process}

\item{theta}{Real value, >= -alpha if alpha is between 0 (inclusive) and 1 (exclusive), or a positive integer multiple of |alpha| if alpha < 0}

\item{n}{Positive integer, number of nodes of the network i.e. order of corresponding graph}

\item{p}{Real value between 0 and 1 inclusive, prior probability of finite regime}

\item{data.seed}{Positive integer, the seed for set.seed()}

\item{dir}{Boolean, is the SBM for directed graphs (TRUE) or directed acyclic graphs (FALSE, default)?}
}
\description{
Wrapper of \code{\link{plot_topo}}
}
\seealso{
\code{\link{obtain_spi}}, \code{\link{read_results}} and \code{\link{compute_bf}} with the same set of arguments
}
