% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{post_mcmc_process}
\alias{post_mcmc_process}
\title{Post-MCMC process results of fitting DAG-SBM}
\usage{
post_mcmc_process(obj, proc_time)
}
\arguments{
\item{obj}{List, object returned from \code{\link{mcmc_dagsbm}}}

\item{proc_time}{A proc_time object, usually from \code{\link{system.time}}}
}
\description{
Post-MCMC process results of fitting DAG-SBM
}
