% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reorder_dense}
\alias{reorder_dense}
\title{Reorder rows and columns of a dense matrix with the same ordering}
\usage{
reorder_dense(Y, sigma)
}
\arguments{
\item{Y}{A dense square matrix}

\item{sigma}{An integer vector that is a permutation of 0, 1, ..., ncol(Y)-1}
}
\description{
Reorder rows and columns of a dense matrix with the same ordering
}
