% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reorder_sparse}
\alias{reorder_sparse}
\title{Reorder rows and columns of a sparse matrix with the same ordering}
\usage{
reorder_sparse(Y, sigma)
}
\arguments{
\item{Y}{A sparse square matrix}

\item{sigma}{An integer vector that is a permutation of 0, 1, ..., ncol(Y)-1}
}
\description{
Reorder rows and columns of a sparse matrix with the same ordering
}
