% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sim_dagsbm}
\alias{sim_dagsbm}
\title{Simulate from DAG-SBM}
\usage{
sim_dagsbm(alpha, theta, n, data.seed, a0, b0, c0, d0)
}
\arguments{
\item{alpha}{Real value less than 1, concentration parameter of the Pitman-Yor process}

\item{theta}{Real value, >= -alpha if alpha is between 0 (inclusive) and 1 (exclusive), or a positive integer multiple of |alpha| if alpha < 0}

\item{n}{Positive integer, number of nodes of the network i.e. order of corresponding graph}

\item{data.seed}{Positive integer, the seed for set.seed()}

\item{a0}{Numeric, hyperparameter for prior of block matrix}

\item{b0}{Numeric, hyperparameter for prior of block matrix}

\item{c0}{Numeric, hyperparameter for prior of degree correction}

\item{d0}{Numeric, hyperparameter for prior of degree correction}
}
\description{
Simulate from DAG-SBM
}
