% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{stop_wrapper}
\alias{stop_wrapper}
\title{Check if object exists, with error message if not}
\usage{
stop_wrapper(string)
}
\arguments{
\item{string}{Scalar string}
}
\description{
Check if object exists, with error message if not
}
