% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ulam}
\alias{ulam}
\title{Shift the position-th element in sigma by distance}
\usage{
ulam(sigma, position, distance)
}
\arguments{
\item{sigma}{A permutation of 0, 1, ..., length(sigma)-1}

\item{position}{integer between 0 and length(sigma)-1 inclusive}

\item{distance}{integer between -length(sigma) and length(sigma) inclusive}
}
\description{
Shift the position-th element in sigma by distance
}
