% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_adj_mat}
\alias{plot_adj_mat}
\title{Plot adjacency matrix with the rows and columns in topological order, using the edgelist and topological ordering}
\usage{
plot_adj_mat(df.edgelist, df.topo, weighted = TRUE)
}
\arguments{
\item{df.edgelist}{Data frame of the edgelist of the graph}

\item{df.topo}{Data frame of the topological ordering of the vertices of the graph}

\item{weighted}{Boolean, is the graph edge-weighted (TRUE) or binary (FALSE)?}
}
\description{
Plot adjacency matrix with the rows and columns in topological order, using the edgelist and topological ordering
}
\keyword{internal}
