% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_adj_psm}
\alias{plot_adj_psm}
\title{Plot adjacency matrix overlaid by similarity (co-clustering) matrix}
\usage{
plot_adj_psm(l, width)
}
\arguments{
\item{l}{asdf}

\item{width}{asdf}
}
\description{
Plot adjacency matrix overlaid by similarity (co-clustering) matrix
}
