% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_network}
\alias{plot_network}
\title{Plot network}
\usage{
plot_network(g, com, mar, seed)
}
\arguments{
\item{g}{asdf}

\item{com}{asdf}

\item{mar}{asdf}

\item{seed}{asdf}
}
\description{
Plot network
}
