% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_topo}
\alias{plot_topo}
\title{Plot posterior of topological ordering}
\usage{
plot_topo(
  df,
  colours = TRUE,
  mean.pos = FALSE,
  low = "yellow",
  true.pos = NULL
)
}
\arguments{
\item{df}{Data frame to be passed to \code{\link{pivot_topo}}}

\item{colours}{Boolean, are colours to be added (default TRUE)?}

\item{mean.pos}{Boolean, are the mean positions to be included as points (default FALSE)?}

\item{low}{String, the colour for the lower end of the scale; ignored if colours is FALSE}

\item{true.pos}{Integer vector, a permutation of 1, 2, ..., length(true.pos), or NULL (default)}
}
\description{
Plot posterior of topological ordering
}
